ENH: move basicSuiteScript stuff to commonPsanaBase, basicSuite not needed
